Move MscParamsHelper to params/detail and fix namespace
